Added debug endpoint to get schedules from ScheduleStore - moved existing code to debug equivalentSchdedules from ContentDebugController to ScheduleDebugController
