+++ atlas-processing/src/main/java/org/atlasapi/system/debug/ContentDebugController.java
@@ -3,14 +3,11 @@ package org.atlasapi.system.debug;
 import java.io.IOException;
 import java.math.BigInteger;
 import java.util.Map;
-import java.util.Set;
 import java.util.concurrent.TimeUnit;
-import java.util.stream.Collectors;
 
 import javax.servlet.http.HttpServletResponse;
 
 import org.atlasapi.AtlasPersistenceModule;
-import org.atlasapi.channel.Channel;
 import org.atlasapi.channel.ChannelResolver;
 import org.atlasapi.content.Container;
 import org.atlasapi.content.Content;
@@ -25,8 +22,6 @@ import org.atlasapi.entity.util.WriteResult;
 import org.atlasapi.equivalence.EquivalenceGraph;
 import org.atlasapi.equivalence.ResolvedEquivalents;
 import org.atlasapi.media.entity.Publisher;
-import org.atlasapi.schedule.EquivalentSchedule;
-import org.atlasapi.schedule.EquivalentScheduleStore;
 import org.atlasapi.system.bootstrap.ContentBootstrapListener;
 import org.atlasapi.system.bootstrap.workers.DirectAndExplicitEquivalenceMigrator;
 import org.atlasapi.system.legacy.LegacyPersistenceModule;
@@ -38,8 +33,6 @@ import com.metabroadcast.common.ids.SubstitutionTableNumberCodec;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.api.client.repackaged.com.google.common.base.Throwables;
-import com.google.common.base.Optional;
-import com.google.common.base.Splitter;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableSet;
 import com.google.common.collect.Iterables;
@@ -50,10 +43,6 @@ import com.google.gson.GsonBuilder;
 import com.google.gson.JsonPrimitive;
 import com.google.gson.JsonSerializer;
 import org.joda.time.DateTime;
-import org.joda.time.Interval;
-import org.joda.time.LocalDate;
-import org.joda.time.format.DateTimeFormatter;
-import org.joda.time.format.ISODateTimeFormat;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -67,9 +56,6 @@ public class ContentDebugController {
 
     private final NumberToShortStringCodec uppercase = new SubstitutionTableNumberCodec();
     private final NumberToShortStringCodec lowercase = SubstitutionTableNumberCodec.lowerCaseOnly();
-    private static final DateTimeFormatter dateParser = ISODateTimeFormat.date();
-    private final ChannelResolver channelResolver;
-    private final EquivalentScheduleStore scheduleStore;
 
     private final Gson gson = new GsonBuilder().registerTypeAdapter(
             DateTime.class,
@@ -90,14 +76,10 @@ public class ContentDebugController {
             LegacyPersistenceModule legacyPersistence,
             AtlasPersistenceModule persistence,
             DirectAndExplicitEquivalenceMigrator equivalenceMigrator,
-            ChannelResolver channelResolver,
-            EquivalentScheduleStore scheduleStore,
             ContentIndex index,
             EsContentTranslator esContentTranslator) {
         this.legacyPersistence = checkNotNull(legacyPersistence);
         this.persistence = checkNotNull(persistence);
-        this.channelResolver = checkNotNull(channelResolver);
-        this.scheduleStore = checkNotNull(scheduleStore);
         this.index = checkNotNull(index);
         this.esContentTranslator = checkNotNull(esContentTranslator);
 
@@ -321,62 +303,7 @@ public class ContentDebugController {
         }
     }
 
-    @RequestMapping("/system/debug/schedule")
-    public void debugEquivalentSchedule(
-            @RequestParam(value = "channel", required = true) String channelId,
-            @RequestParam(value = "source", required = true) String sourceKey,
-            @RequestParam(value = "selectedSources", required = false) String selectedSourcesKeys,
-            @RequestParam(value = "day", required = true) String day,
-            final HttpServletResponse response
-    ) throws IOException {
-        try {
 
-            Publisher source = Publisher.fromKey(sourceKey).requireValue();
-
-            Set<Publisher> selectedSources = null;
-
-            if (selectedSourcesKeys != null) {
-                Splitter.on(",")
-                        .splitToList(selectedSourcesKeys)
-                        .stream()
-                        .map(key -> Publisher.fromKey(key).requireValue())
-                        .collect(Collectors.toSet());
-            } else {
-                selectedSources = Publisher.all();
-            }
-
-            LocalDate date;
-            try {
-                date = dateParser.parseLocalDate(day);
-            } catch (IllegalArgumentException iae) {
-                response.setStatus(400);
-                response.getWriter().write("Failed to parse " + day + ", expected yyyy-MM-dd");
-                return;
-            }
-
-            Optional<Channel> channel = resolve(channelId);
-            if (!channel.isPresent()) {
-                response.setStatus(404);
-                response.getWriter().write("Unknown channel " + channelId);
-            }
-
-            EquivalentSchedule equivalentSchedule = Futures.get(scheduleStore.resolveSchedules(
-                    ImmutableSet.of(channel.get()),
-                    new Interval(
-                            date.toDateTimeAtStartOfDay(),
-                            date.plusDays(1).toDateTimeAtStartOfDay()
-                    ),
-                    source,
-                    selectedSources
-            ), Exception.class);
-
-            gson.toJson(equivalentSchedule, response.getWriter());
-            response.setStatus(200);
-            response.flushBuffer();
-        } catch (Throwable t) {
-            t.printStackTrace(response.getWriter());
-        }
-    }
 
     private Content resolveLegacyContent(Long id) {
         return Iterables.getOnlyElement(
@@ -385,20 +312,4 @@ public class ContentDebugController {
         );
     }
 
-    private Optional<Channel> resolve(String channelId) throws Exception {
-        Id cid = Id.valueOf(lowercase.decode(channelId));
-        ListenableFuture<Resolved<Channel>> channelFuture = channelResolver.resolveIds(ImmutableList
-                .of(cid));
-        Resolved<Channel> resolvedChannel = Futures.get(
-                channelFuture,
-                1,
-                TimeUnit.MINUTES,
-                Exception.class
-        );
-
-        if (resolvedChannel.getResources().isEmpty()) {
-            return Optional.absent();
-        }
-        return Optional.of(Iterables.getOnlyElement(resolvedChannel.getResources()));
-    }
 }
